Renamed dataset collection response objects
