+++ src/main/java/com/github/jmchilton/blend4j/galaxy/HistoriesClientImpl.java
@@ -14,7 +14,7 @@ import com.github.jmchilton.blend4j.galaxy.beans.HistoryExport;
 import com.github.jmchilton.blend4j.galaxy.beans.collection.request.CollectionElement;
 import com.github.jmchilton.blend4j.galaxy.beans.collection.request.CollectionDescription;
 import com.github.jmchilton.blend4j.galaxy.beans.collection.request.HistoryDatasetElement;
-import com.github.jmchilton.blend4j.galaxy.beans.collection.response.DatasetCollectionResponse;
+import com.github.jmchilton.blend4j.galaxy.beans.collection.response.CollectionResponse;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
 
@@ -80,10 +80,10 @@ class HistoriesClientImpl extends Client implements HistoriesClient {
   }
 
   @Override
-  public DatasetCollectionResponse showDatasetCollection(String historyId,
+  public CollectionResponse showDatasetCollection(String historyId,
       String datasetCollectionId) {
     return getWebResourceContents(historyId).path("dataset_collections").
-        path(datasetCollectionId).get(DatasetCollectionResponse.class);
+        path(datasetCollectionId).get(CollectionResponse.class);
   }
   
   @Override
@@ -94,12 +94,12 @@ class HistoriesClientImpl extends Client implements HistoriesClient {
   }
 
   @Override
-  public DatasetCollectionResponse createDatasetCollection(String historyId,
+  public CollectionResponse createDatasetCollection(String historyId,
       CollectionDescription collectionDescription) {
     ClientResponse response = createDatasetCollectionRequest(historyId, collectionDescription);
     
     if (response.getStatus() == 200) {
-      return response.getEntity(DatasetCollectionResponse.class);
+      return response.getEntity(CollectionResponse.class);
     } else {
       throw new RuntimeException("Error creating dataset collection, status=" + response.getStatus() +
           " returned=" + response.getEntity(String.class));
