+++ src/main/java/com/github/jmchilton/blend4j/galaxy/beans/collection/response/ElementResponse.java
@@ -1,0 +1,8 @@
+package com.github.jmchilton.blend4j.galaxy.beans.collection.response;
+
+/**
+ * Interface defining an element within a dataset collection.
+ */
+public interface ElementResponse {
+
+}
