--- src/main/java/com/github/jmchilton/blend4j/galaxy/beans/collection/response/ResponseObject.java
@@ -1,8 +1,0 @@
-package com.github.jmchilton.blend4j.galaxy.beans.collection.response;
-
-/**
- * Interface defining a response to a dataset collection request.
- */
-public interface ResponseObject {
-
-}
